backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pytesseract
Pillow
pdfminer.six
scikit-learn
nltk
sqlalchemy
alembic
python-dotenv
psycopg2-binary
pgvector
email-validator
spacy>=3.0.0
transformers>=4.30.0
torch>=2.0.0
accelerate
sentence-transformers>=2.2.0
